Trekk felles klasser for V3 og V4 ut i egen fil
