+++ src/main/kotlin/no/nav/familie/klage/fagsak/domain/FagsakDomain.kt
@@ -1,7 +1,7 @@
 package no.nav.familie.klage.fagsak.domain
 
 import no.nav.familie.klage.felles.domain.Sporbar
-import no.nav.familie.klage.kabal.Ytelse
+import no.nav.familie.klage.kabal.domain.Ytelse
 import no.nav.familie.kontrakter.felles.klage.Fagsystem
 import no.nav.familie.kontrakter.felles.klage.Stønadstype
 import org.springframework.data.annotation.Id
